converting data.js to ts to fix errors works on #19
