+++ public/index.html
@@ -29,8 +29,7 @@
     }));
   </script>
 
-  <script type="text/javascript" src="code/wet.js"></script>
-  <script type="text/javascript" src="code/data.js"></script>
+  <script type="text/javascript" src="helper/wet.js"></script>
 </head>
 
   <body vocab="http://schema.org" typeof="WebPage">
